Cleanup: consolidate files
