+++ admin/client/AdminApp.tsx
@@ -31,7 +31,7 @@ import { LoadingBlocker, Modal } from "./Forms"
 import { AdminAppContext } from "./AdminAppContext"
 import { Base64 } from "js-base64"
 import { ExplorerCreatePage } from "dataExplorer/admin/ExplorerCreatePage"
-import { ExplorersIndexPage } from "dataExplorer/ExplorersIndexPage"
+import { ExplorersIndexPage } from "dataExplorer/admin/ExplorersIndexPage"
 
 @observer
 class AdminErrorMessage extends React.Component<{ admin: Admin }> {
