+++ dataExplorer/client/DataExplorerShell.tsx
@@ -1,28 +1,20 @@
-import React from "react"
-import { observer } from "mobx-react"
-import { CommandPalette, Command } from "charts/CommandPalette"
-import classNames from "classnames"
-import { computed, action, observable, when, reaction } from "mobx"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { Bounds } from "charts/Bounds"
-import { faChartLine } from "@fortawesome/free-solid-svg-icons/faChartLine"
-import { ChartView } from "charts/ChartView"
-import { CountryPicker } from "charts/CountryPicker"
-import { ExplorerControlBar } from "dataExplorer/client/ExplorerControls"
-import { ChartConfig, ChartConfigProps } from "charts/ChartConfig"
-import { throttle, uniq } from "charts/Util"
-import { SwitcherOptions } from "./SwitcherOptions"
-import { ExplorerControlPanel } from "dataExplorer/client/ExplorerControls"
-import {
-    ChartQueryParams,
-    ExtendedChartUrl,
-    EntityUrlBuilder
-} from "charts/ChartUrl"
-import ReactDOM from "react-dom"
-import { UrlBinder } from "charts/UrlBinder"
+import { computed, action, observable } from "mobx"
 import { QueryParams } from "utils/client/url"
+import { EntityUrlBuilder } from "charts/ChartUrl"
+import { observer } from "mobx-react"
+import React from "react"
+import { ChartConfig } from "charts/ChartConfig"
+import { Command, CommandPalette } from "charts/CommandPalette"
+import { Bounds } from "charts/Bounds"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+import { faChartLine } from "@fortawesome/free-solid-svg-icons/faChartLine"
+import { CountryPicker } from "charts/CountryPicker"
+import { ExplorerControlBar } from "./ExplorerControls"
+import { throttle } from "lodash"
+import classNames from "classnames"
+import { ChartView } from "charts/ChartView"
 
-class DataExplorerQueryParams {
+export class DataExplorerQueryParams {
     hideControls: boolean = false
     selectedCountryCodes: Set<string> = new Set<string>()
 
@@ -36,11 +28,9 @@ class DataExplorerQueryParams {
     }
 }
 
-declare type chartId = number
-
 // TODO: Migrate CovidExplorer to use this class as well
 @observer
-class DataExplorerShell extends React.Component<{
+export class DataExplorerShell extends React.Component<{
     explorerName: string
     controlPanels: JSX.Element[]
     chart: ChartConfig
@@ -225,162 +215,6 @@ class DataExplorerShell extends React.Component<{
                     </div>
                 </div>
             </>
-        )
-    }
-}
-
-export interface SwitcherDataExplorerProps {
-    chartConfigs: Map<chartId, ChartConfigProps>
-    switcher: SwitcherOptions
-    explorerNamespace: string
-    explorerTitle: string
-    bindToWindow: boolean
-}
-
-@observer
-export class SwitcherDataExplorer extends React.Component<
-    SwitcherDataExplorerProps
-> {
-    static bootstrap(
-        containerNode: HTMLElement,
-        chartConfigs: ChartConfigProps[],
-        switcherCode: string,
-        title: string,
-        bindToWindow: boolean
-    ) {
-        const switcher = new SwitcherOptions(switcherCode, "")
-
-        const chartConfigsMap: Map<number, ChartConfigProps> = new Map()
-        chartConfigs.forEach(config => chartConfigsMap.set(config.id!, config))
-
-        return ReactDOM.render(
-            <SwitcherDataExplorer
-                chartConfigs={chartConfigsMap}
-                explorerNamespace="explorer"
-                explorerTitle={title}
-                switcher={switcher}
-                bindToWindow={bindToWindow}
-            />,
-            containerNode
-        )
-    }
-
-    urlBinding?: UrlBinder
-
-    bindToWindow() {
-        const url = new ExtendedChartUrl(this._chart!.url, [
-            this.props.switcher.toParams(),
-            this.userOptions.toParams
-        ])
-
-        if (this.urlBinding) this.urlBinding.unbindFromWindow()
-        else this.urlBinding = new UrlBinder()
-
-        this.urlBinding.bindToWindow(url)
-    }
-
-    componentWillMount() {
-        reaction(() => this.props.switcher.chartId, this.updateChart, {
-            fireImmediately: true
-        })
-
-        const win = window as any
-        win.switcherDataExplorer = this
-    }
-
-    @action.bound updateChart() {
-        const newId: number = this.props.switcher.chartId
-        if (newId === this.lastId) return
-
-        const params = this.changedParams
-        const props =
-            this.props.chartConfigs.get(newId) || new ChartConfigProps()
-
-        this._chart = new ChartConfig(props)
-        this._chart.url.populateFromQueryParams(params)
-
-        if (this.props.bindToWindow) this.bindToWindow()
-
-        when(
-            () => this._chart!.isReady,
-            () => {
-                this.availableEntities = uniq([
-                    ...this.availableEntities,
-                    ...this._chart!.table.availableEntities
-                ]).sort()
-
-                this._chart!.props.selectedData = this.selectedData
-            }
-        )
-
-        this.lastId = newId
-    }
-
-    private get selectedData() {
-        const table = this._chart!.table
-        const countryCodeMap = table.entityCodeToNameMap
-        const entityIdMap = table.entityNameToIdMap
-        return Array.from(this.userOptions.selectedCountryCodes)
-            .map(code => countryCodeMap.get(code))
-            .filter(i => i)
-            .map(countryOption => {
-                return {
-                    index: 0,
-                    entityId: countryOption
-                        ? entityIdMap.get(countryOption)!
-                        : 0
-                }
-            })
-    }
-
-    get changedParams(): ChartQueryParams {
-        return this._chart?.url.params || {}
-    }
-
-    @observable private _chart?: ChartConfig = undefined
-    private lastId = 0
-
-    @observable availableEntities: string[] = []
-
-    get panels() {
-        return this.props.switcher.groups.map(group => (
-            <ExplorerControlPanel
-                key={group.title}
-                title={group.title}
-                explorerName={this.props.explorerNamespace}
-                name={group.title}
-                options={group.options}
-                isCheckbox={group.isCheckbox}
-                onChange={value => {
-                    this.props.switcher.setValue(group.title, value)
-                }}
-            />
-        ))
-    }
-
-    get header() {
-        return (
-            <>
-                <div></div>
-                <div className="ExplorerTitle">{this.props.explorerTitle}</div>
-                <div className="ExplorerLastUpdated"></div>
-            </>
-        )
-    }
-
-    @observable userOptions = new DataExplorerQueryParams()
-
-    render() {
-        return (
-            <DataExplorerShell
-                headerElement={this.header}
-                controlPanels={this.panels}
-                explorerName={this.props.explorerNamespace}
-                availableEntities={this.availableEntities}
-                chart={this._chart!}
-                params={this.userOptions}
-                isEmbed={false}
-            />
         )
     }
 }
